+++ src/app/(dashboard)/dashboard/providers/[id]/ProviderDetailPageClient.tsx
@@ -93,9 +93,10 @@ import ConnectionRow, {
 } from "./components/ConnectionRow";
 import ModelCompatPopover from "./components/ModelCompatPopover";
 import SiliconFlowEndpointModal from "./components/SiliconFlowEndpointModal";
-import { CC_COMPATIBLE_DEFAULT_CHAT_PATH } from "./providerDetailConstants";
 // Phase 1k extractions — Issue #3501
 import { useModelImportHandlers } from "./hooks/useModelImportHandlers";
+// Phase 1s extractions — Issue #3501
+import { useApiKeySave } from "./hooks/useApiKeySave";
 import ImportProgressModal from "./components/ImportProgressModal";
 // Phase 1l extractions — Issue #3501
 import { useModelVisibilityHandlers } from "./hooks/useModelVisibilityHandlers";
@@ -119,6 +120,11 @@ import {
   type LocalProviderMetadata,
   // CommandCodeAuthFlowState moved to hooks/useCommandCodeAuth.ts (Phase 1h)
   // CompatByProtocolMap, CompatModelRow, CompatModelMap → hooks/useModelVisibilityHandlers.ts (Phase 1l)
+  // Phase 1s: pure helpers extracted from god-component closures
+  getApiLabel,
+  getApiDefaultPath,
+  getApiPath,
+  getHeaderIconProviderId,
 } from "./providerPageHelpers";
 // CODEX_GLOBAL_SERVICE_MODE_VALUES, getCodexServiceTierLabel, normalizeCodexLimitPolicy
 // moved to hooks/useProviderSettings.ts + hooks/useProviderConnections.ts (Phase 1f)
@@ -128,6 +134,15 @@ import ModelRow, { ModelVisibilityToolbar } from "./components/ModelRow";
 import PassthroughModelsSection from "./components/PassthroughModelsSection";
 import CustomModelsSection from "./components/CustomModelsSection";
 import CompatibleModelsSection from "./components/CompatibleModelsSection";
+import ConnectionsListPanel from "./components/ConnectionsListPanel";
+// Phase 1o extractions — Issue #3501
+import ConnectionsHeaderToolbar from "./components/ConnectionsHeaderToolbar";
+// Phase 1p extractions — Issue #3501
+import ZedImportCard from "./components/ZedImportCard";
+// Phase 1q extractions — Issue #3501
+import BatchTestResultsModal from "./components/BatchTestResultsModal";
+// Phase 1r extractions — Issue #3501
+import { AdaptaTutorialModal } from "./components/AdaptaTutorialModal";
 // recordToHeaderRows moved to components/ModelCompatPopover.tsx (Phase 1d)
 // buildCompatMap, isModelHidden*, effectiveNormalize/Preserve*, anyNormalize/NoPreserveCompatBadge
 // moved to providerPageHelpers.ts + hook useModelCompatState (Phase 1e)
@@ -165,11 +180,6 @@ export default function ProviderDetailPageClient() {
   const [showTutorialModal, setShowTutorialModal] = useState(false);
   const [selectedConnection, setSelectedConnection] = useState(null);
   const [proxyTarget, setProxyTarget] = useState(null);
-  const [importingZed, setImportingZed] = useState(false);
-  const [showZedManual, setShowZedManual] = useState(false);
-  const [zedManualProvider, setZedManualProvider] = useState("openai");
-  const [zedManualToken, setZedManualToken] = useState("");
-  const [importingZedManual, setImportingZedManual] = useState(false);
   const [importCodexModalOpen, setImportCodexModalOpen] = useState(false);
   const [codexCliGuideOpen, setCodexCliGuideOpen] = useState(false);
   const [importClaudeModalOpen, setImportClaudeModalOpen] = useState(false);
@@ -379,50 +389,6 @@ export default function ProviderDetailPageClient() {
     providerStorageAlias,
   });
 
-  const getApiLabel = () => {
-    if (isAnthropicProtocolCompatible) return t("messagesApi");
-    const type = providerNode?.apiType;
-    switch (type) {
-      case "responses":
-        return t("responsesApi");
-      case "embeddings":
-        return t("embeddings");
-      case "audio-transcriptions":
-        return t("audioTranscriptions");
-      case "audio-speech":
-        return t("audioSpeech");
-      case "images-generations":
-        return t("imagesGenerations");
-      default:
-        return t("chatCompletions");
-    }
-  };
-
-  const getApiDefaultPath = () => {
-    if (isCcCompatible) return CC_COMPATIBLE_DEFAULT_CHAT_PATH;
-    if (isAnthropicCompatible) return "/messages";
-    const type = providerNode?.apiType;
-    switch (type) {
-      case "responses":
-        return "/responses";
-      case "embeddings":
-        return "/embeddings";
-      case "audio-transcriptions":
-        return "/audio/transcriptions";
-      case "audio-speech":
-        return "/audio/speech";
-      case "images-generations":
-        return "/images/generations";
-      default:
-        return "/chat/completions";
-    }
-  };
-
-  const getApiPath = () => {
-    const defaultPath = getApiDefaultPath();
-    return (providerNode?.chatPath || defaultPath).replace(/^\//, "");
-  };
-
   // fetchAliases, handleSetAlias, handleDeleteAlias → hooks/useProviderModels.ts (Phase 1f)
   // fetchProviderModelMeta, fetchProxyConfig, fetchConnections → hooks/useProviderConnections.ts + useProviderModels.ts (Phase 1f)
   // loadCodexSettings, loadClaudeRoutingSettings → hooks/useProviderSettings.ts (Phase 1f)
@@ -458,63 +424,6 @@ export default function ProviderDetailPageClient() {
       console.log("Error updating provider node:", error);
     }
   };
-
-  const handleZedImport = useCallback(async () => {
-    if (importingZed) return;
-    setImportingZed(true);
-    try {
-      const res = await fetch("/api/providers/zed/import", { method: "POST" });
-      const data = await res.json();
-      if (!res.ok || !data.success) {
-        if (data.zedDockerEnvironment) {
-          setShowZedManual(true);
-        }
-        notify.error(data.error || "Zed import failed");
-      } else if (!data.count) {
-        const found = data.credentials?.length ?? 0;
-        if (found === 0) {
-          notify.info("No Zed credentials found in keychain");
-        } else {
-          notify.info(
-            `Found ${found} keychain credential(s), but none matched supported providers`
-          );
-        }
-      } else {
-        notify.success(
-          `Imported ${data.count} credential(s) from Zed for ${data.providers?.length ?? 0} provider(s)`
-        );
-        await fetchConnections();
-      }
-    } catch (e: any) {
-      notify.error(e?.message || "Zed import failed");
-    } finally {
-      setImportingZed(false);
-    }
-  }, [importingZed, notify, fetchConnections]);
-
-  const handleZedManualImport = useCallback(async () => {
-    if (importingZedManual || !zedManualToken.trim()) return;
-    setImportingZedManual(true);
-    try {
-      const res = await fetch("/api/providers/zed/manual-import", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ provider: zedManualProvider, token: zedManualToken.trim() }),
-      });
-      const data = await res.json();
-      if (!res.ok || !data.success) {
-        notify.error(data.error?.message ?? data.error ?? "Manual import failed");
-      } else {
-        notify.success(`Imported ${zedManualProvider} token from Zed`);
-        setZedManualToken("");
-        await fetchConnections();
-      }
-    } catch (e: any) {
-      notify.error(e?.message || "Manual import failed");
-    } finally {
-      setImportingZedManual(false);
-    }
-  }, [importingZedManual, zedManualProvider, zedManualToken, notify, fetchConnections]);
 
   // loadCodexSettings, loadClaudeRoutingSettings → hooks/useProviderSettings.ts (Phase 1f)
   // loadConnProxies → hooks/useProviderConnections.ts (Phase 1f)
@@ -587,98 +496,18 @@ export default function ProviderDetailPageClient() {
     notify,
   });
 
-  const handleSaveApiKey = async (formData) => {
-    try {
-      const res = await fetch("/api/providers", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ provider: providerId, ...formData }),
-      });
-      if (res.ok) {
-        const connectionData = await res.json();
-        const newConnection = connectionData?.connection;
-        await fetchConnections();
-        setShowAddApiKeyModal(false);
-        setSiliconFlowInitialBaseUrl(undefined);
-
-        // Universal: sync models from the provider endpoint on every new connection
-        // (was previously Gemini-only). Do NOT re-introduce a providerId guard here.
-        if (newConnection?.id) {
-          setShowImportModal(true);
-          setImportProgress({
-            current: 0,
-            total: 0,
-            phase: "fetching",
-            status: t("fetchingModels"),
-            logs: [],
-            error: "",
-            importedCount: 0,
-          });
-
-          try {
-            const syncRes = await fetch(`/api/providers/${newConnection.id}/sync-models`, {
-              method: "POST",
-              signal: AbortSignal.timeout(30_000), // 30s timeout — model sync shouldn't hang
-            });
-            const syncData = await syncRes.json();
-
-            if (!syncRes.ok || syncData.error) {
-              setImportProgress((prev) => ({
-                ...prev,
-                phase: "error",
-                status: t("failedFetchModels"),
-                error: syncData.error?.message || syncData.error || t("failedImportModels"),
-              }));
-              return null;
-            }
-
-            const syncedCount = syncData.syncedModels || 0;
-            const availableCount =
-              typeof syncData.availableModelsCount === "number"
-                ? syncData.availableModelsCount
-                : Array.isArray(syncData.models)
-                  ? syncData.models.length
-                  : syncedCount;
-            const syncedModelList: Array<{ id: string; name?: string }> = syncData.models || [];
-            const logs: string[] = [];
-            if (syncedModelList.length > 0) {
-              logs.push(`✓ ${availableCount} models available`);
-              logs.push("");
-              for (const m of syncedModelList) {
-                logs.push(`  ${m.name || m.id}`);
-              }
-            }
-
-            setImportProgress((prev) => ({
-              ...prev,
-              phase: "done",
-              status: t("modelsImported", { count: availableCount }),
-              total: availableCount,
-              current: availableCount,
-              importedCount: availableCount,
-              logs,
-            }));
-
-            await fetchProviderModelMeta();
-          } catch (syncError) {
-            setImportProgress((prev) => ({
-              ...prev,
-              phase: "error",
-              status: t("failedFetchModels"),
-              error: String(syncError),
-            }));
-          }
-        }
-        return null;
-      }
-      const data = await res.json().catch(() => ({}));
-      const errorMsg = data.error?.message || data.error || t("failedSaveConnection");
-      return errorMsg;
-    } catch (error) {
-      console.log("Error saving connection:", error);
-      return t("failedSaveConnectionRetry");
-    }
-  };
+  // Phase 1s: handleSaveApiKey extracted to hooks/useApiKeySave.ts
+  const { handleSaveApiKey } = useApiKeySave({
+    providerId,
+    fetchConnections,
+    fetchProviderModelMeta,
+    setImportProgress,
+    setShowImportModal,
+    setShowAddApiKeyModal,
+    setSiliconFlowInitialBaseUrl,
+    notify,
+    t,
+  });
 
   const handleUpdateConnection = async (formData) => {
     try {
@@ -828,17 +657,6 @@ export default function ProviderDetailPageClient() {
     );
   }
 
-  // OpenAI/Anthropic compatible providers use their specialized pseudo-provider icons.
-  const getHeaderIconProviderId = () => {
-    if (isOpenAICompatible && providerInfo.apiType) {
-      return providerInfo.apiType === "responses" ? "oai-r" : "oai-cc";
-    }
-    if (isAnthropicProtocolCompatible) {
-      return "anthropic-m";
-    }
-    return providerInfo.id;
-  };
-
   return (
     <div className="flex flex-col gap-8">
       {/* Header */}
@@ -855,7 +673,7 @@ export default function ProviderDetailPageClient() {
             className="rounded-lg flex items-center justify-center"
             style={{ backgroundColor: `${providerInfo.color}15` }}
           >
-            <ProviderIcon providerId={getHeaderIconProviderId()} size={48} type="color" />
+            <ProviderIcon providerId={getHeaderIconProviderId(isOpenAICompatible, isAnthropicProtocolCompatible, providerInfo.id, providerInfo.apiType)} size={48} type="color" />
           </div>
           <div>
             {providerInfo.website ? (
@@ -891,91 +709,7 @@ export default function ProviderDetailPageClient() {
         </div>
       </div>
 
-      {providerId === "zed" && (
-        <>
-          <Card>
-            <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-              <div className="flex-1 min-w-0">
-                <h2 className="text-lg font-semibold flex items-center gap-2">
-                  <span className="material-symbols-outlined text-[20px]">download</span>
-                  Import from Zed Keychain
-                </h2>
-                <p className="text-sm text-text-muted mt-1">
-                  Discover AI provider credentials (OpenAI, Anthropic, Google, Mistral, xAI) that
-                  Zed IDE stored in the OS keychain and import them as connections. Requires Zed IDE
-                  installed on this machine.
-                </p>
-              </div>
-              <Button
-                size="sm"
-                variant="secondary"
-                icon={importingZed ? "sync" : "download"}
-                onClick={handleZedImport}
-                disabled={importingZed}
-              >
-                {importingZed ? "Importing…" : "Import from Zed"}
-              </Button>
-            </div>
-          </Card>
-          <Card>
-            <div className="flex flex-col gap-3">
-              <button
-                className="flex items-center justify-between w-full text-left"
-                onClick={() => setShowZedManual((v) => !v)}
-              >
-                <h2 className="text-lg font-semibold flex items-center gap-2">
-                  <span className="material-symbols-outlined text-[20px]">edit</span>
-                  Manual Token Import
-                </h2>
-                <span className="material-symbols-outlined text-[18px] text-text-muted">
-                  {showZedManual ? "expand_less" : "expand_more"}
-                </span>
-              </button>
-              {showZedManual && (
-                <div className="flex flex-col gap-3 mt-1">
-                  <p className="text-sm text-text-muted">
-                    Use this when OmniRoute runs in Docker or the keychain is unavailable. Paste the
-                    API key that Zed stored under{" "}
-                    <code className="font-mono text-xs">~/.config/zed/settings.json</code> or copy
-                    it from the Zed AI settings panel.
-                  </p>
-                  <div className="flex gap-2 flex-col sm:flex-row">
-                    <select
-                      className="input input-sm"
-                      value={zedManualProvider}
-                      onChange={(e) => setZedManualProvider(e.target.value)}
-                    >
-                      <option value="openai">OpenAI</option>
-                      <option value="anthropic">Anthropic</option>
-                      <option value="google">Google</option>
-                      <option value="mistral">Mistral</option>
-                      <option value="xai">xAI</option>
-                      <option value="openrouter">OpenRouter</option>
-                      <option value="deepseek">DeepSeek</option>
-                    </select>
-                    <input
-                      type="password"
-                      className="input input-sm flex-1"
-                      placeholder="Paste API key…"
-                      value={zedManualToken}
-                      onChange={(e) => setZedManualToken(e.target.value)}
-                    />
-                    <Button
-                      size="sm"
-                      variant="secondary"
-                      icon={importingZedManual ? "sync" : "upload"}
-                      onClick={handleZedManualImport}
-                      disabled={importingZedManual || !zedManualToken.trim()}
-                    >
-                      {importingZedManual ? "Saving…" : "Import"}
-                    </Button>
-                  </div>
-                </div>
-              )}
-            </div>
-          </Card>
-        </>
-      )}
+      {providerId === "zed" && <ZedImportCard fetchConnections={fetchConnections} notify={notify} />}
 
       {isCompatible && providerNode && (
         <Card>
@@ -989,7 +723,7 @@ export default function ProviderDetailPageClient() {
                     : t("openaiCompatibleDetails")}
               </h2>
               <p className="text-sm text-text-muted">
-                {getApiLabel()} · {(providerNode.baseUrl || "").replace(/\/$/, "")}/{getApiPath()}
+                {getApiLabel(t, isAnthropicProtocolCompatible, providerNode?.apiType)} · {(providerNode.baseUrl || "").replace(/\/$/, "")}/{getApiPath(isCcCompatible, isAnthropicCompatible, providerNode?.apiType, providerNode?.chatPath)}
               </p>
             </div>
             <div className="flex flex-wrap items-center gap-2">
@@ -1071,290 +805,48 @@ export default function ProviderDetailPageClient() {
         providerId !== "opencode" && <NoAuthProviderCard />}
       {!isUpstreamProxyProvider && !isFreeNoAuth && (
         <Card>
-          <div className="mb-4 flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-            <div className="flex min-w-0 flex-wrap items-center gap-2">
-              <h2 className="text-lg font-semibold">{t("connections")}</h2>
-              {providerId === "claude" && (
-                <div
-                  className="inline-flex items-center gap-2 rounded-lg border border-orange-500/20 bg-orange-500/5 px-2 py-1 text-xs font-medium text-text-muted"
-                  title={providerText(
-                    t,
-                    "preferClaudeCodeForUnprefixedClaudeModelsTooltip",
-                    "Route bare claude-* model IDs from Claude Code clients through the Claude Code account instead of asking for a provider prefix."
-                  )}
-                >
-                  <span className="material-symbols-outlined text-[14px] text-orange-500">
-                    alt_route
-                  </span>
-                  <span>
-                    {providerText(
-                      t,
-                      "preferClaudeCodeForUnprefixedClaudeModelsLabel",
-                      "Claude Code default"
-                    )}
-                  </span>
-                  <Toggle
-                    size="sm"
-                    checked={preferClaudeCodeForUnprefixedClaudeModels}
-                    onChange={handleToggleClaudeRoutingPreference}
-                    disabled={savingClaudeRoutingPreference || !claudeRoutingSettingsLoaded}
-                    ariaLabel={providerText(
-                      t,
-                      "preferClaudeCodeForUnprefixedClaudeModelsAria",
-                      "Prefer Claude Code for unprefixed Claude models"
-                    )}
-                    title={
-                      preferClaudeCodeForUnprefixedClaudeModels
-                        ? providerText(
-                            t,
-                            "preferClaudeCodeForUnprefixedClaudeModelsDisable",
-                            "Disable Claude Code preference for bare claude-* model IDs"
-                          )
-                        : providerText(
-                            t,
-                            "preferClaudeCodeForUnprefixedClaudeModelsEnable",
-                            "Enable Claude Code preference for bare claude-* model IDs"
-                          )
-                    }
-                  />
-                  <span className="text-[11px] text-text-muted/70">
-                    {preferClaudeCodeForUnprefixedClaudeModels
-                      ? providerText(t, "toggleOnShort", "On")
-                      : providerText(t, "toggleOffShort", "Off")}
-                  </span>
-                  {claudeRoutingSettingsLoadError ? (
-                    <button
-                      type="button"
-                      onClick={() => void loadClaudeRoutingSettings()}
-                      className="rounded border border-orange-500/30 px-2 py-0.5 text-[11px] font-medium text-orange-600 hover:bg-orange-500/10 dark:text-orange-300"
-                      title={claudeRoutingSettingsLoadError}
-                    >
-                      {providerText(t, "retry", "Retry")}
-                    </button>
-                  ) : null}
-                </div>
-              )}
-              {providerId === "codex" && (
-                <div
-                  className="inline-flex items-center gap-2 rounded-lg border border-sky-500/20 bg-sky-500/5 px-2 py-1 text-xs font-medium text-text-muted"
-                  title={providerText(
-                    t,
-                    "providerDetailServiceModeTooltip",
-                    "Set a global Codex service mode, or leave accounts on their individual service-tier setting."
-                  )}
-                >
-                  <span>
-                    {providerText(t, "providerDetailServiceModeLabel", "Global service mode:")}
-                  </span>
-                  <select
-                    value={codexGlobalServiceMode}
-                    onChange={(event) =>
-                      handleChangeCodexGlobalServiceMode(
-                        event.target.value as CodexGlobalServiceMode
-                      )
-                    }
-                    disabled={savingCodexGlobalServiceMode || !codexSettingsLoaded}
-                    aria-label="Global Codex service mode"
-                    className="rounded-md border border-border bg-bg px-2 py-1 text-xs text-text-main outline-none transition-colors focus:border-primary disabled:opacity-60"
-                  >
-                    {codexGlobalServiceModeOptions.map((option) => (
-                      <option key={option.value} value={option.value}>
-                        {option.label}
-                      </option>
-                    ))}
-                  </select>
-                  {codexSettingsLoadError ? (
-                    <button
-                      type="button"
-                      onClick={() => void loadCodexSettings()}
-                      className="rounded border border-sky-500/30 px-2 py-0.5 text-[11px] font-medium text-sky-600 hover:bg-sky-500/10 dark:text-sky-300"
-                      title={codexSettingsLoadError}
-                    >
-                      {providerText(t, "retry", "Retry")}
-                    </button>
-                  ) : null}
-                </div>
-              )}
-              {/* Provider-level proxy indicator/button */}
-              <button
-                onClick={() =>
-                  setProxyTarget({
-                    level: "provider",
-                    id: providerId,
-                    label: providerInfo?.name || providerId,
-                  })
-                }
-                className={`inline-flex items-center gap-1 px-2 py-1 rounded text-xs font-medium transition-all ${
-                  proxyConfig?.providers?.[providerId]
-                    ? "bg-amber-500/15 text-amber-500 hover:bg-amber-500/25"
-                    : "bg-black/[0.03] dark:bg-white/[0.03] text-text-muted/50 hover:text-text-muted hover:bg-black/[0.06] dark:hover:bg-white/[0.06]"
-                }`}
-                title={
-                  proxyConfig?.providers?.[providerId]
-                    ? t("providerProxyTitleConfigured", {
-                        host: proxyConfig.providers[providerId].host || t("configured"),
-                      })
-                    : t("providerProxyConfigureHint")
-                }
-              >
-                <span className="material-symbols-outlined text-[14px]">vpn_lock</span>
-                {proxyConfig?.providers?.[providerId]
-                  ? proxyConfig.providers[providerId].host || t("providerProxy")
-                  : t("providerProxy")}
-              </button>
-            </div>
-            <div className="flex shrink-0 flex-wrap items-center justify-end gap-2">
-              {connections.length > 0 && (
-                <button
-                  onClick={() => handleDistributeProxies()}
-                  disabled={distributingProxies || batchTesting || !!retestingId}
-                  className={`flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-medium border transition-colors ${
-                    distributingProxies
-                      ? "bg-primary/20 border-primary/40 text-primary animate-pulse"
-                      : "bg-bg-subtle border-border text-text-muted hover:text-text-primary hover:border-primary/40"
-                  }`}
-                  title={t("distributeProxies")}
-                  aria-label={t("distributeProxies")}
-                >
-                  <span className="material-symbols-outlined text-[14px]">
-                    {distributingProxies ? "sync" : "swap_horiz"}
-                  </span>
-                  {distributingProxies ? t("distributing") : t("distributeProxies")}
-                </button>
-              )}
-              {connections.length > 1 && (
-                <button
-                  onClick={handleBatchTestAll}
-                  disabled={batchTesting || batchRetesting || !!retestingId}
-                  className={`flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-medium border transition-colors ${
-                    batchTesting
-                      ? "bg-primary/20 border-primary/40 text-primary animate-pulse"
-                      : "bg-bg-subtle border-border text-text-muted hover:text-text-primary hover:border-primary/40"
-                  }`}
-                  title={t("testAll")}
-                  aria-label={t("testAll")}
-                >
-                  <span className="material-symbols-outlined text-[14px]">
-                    {batchTesting ? "sync" : "play_arrow"}
-                  </span>
-                  {batchTesting ? t("testing") : t("testAll")}
-                </button>
-              )}
-              {!isCompatible ? (
-                <>
-                  {isCommandCode ? (
-                    <>
-                      <Button
-                        size="sm"
-                        icon="open_in_new"
-                        loading={
-                          commandCodeAuthState.phase === "starting" ||
-                          commandCodeAuthState.phase === "polling" ||
-                          commandCodeAuthState.phase === "applying"
-                        }
-                        onClick={() => gateConnectionFlow(handleOpenCommandCodeConnect)}
-                      >
-                        Connect
-                      </Button>
-                      <Button
-                        size="sm"
-                        variant="secondary"
-                        icon="add"
-                        onClick={() => gateConnectionFlow(openApiKeyAddFlow)}
-                      >
-                        Manual API key
-                      </Button>
-                    </>
-                  ) : (
-                    <>
-                      <Button
-                        size="sm"
-                        icon="add"
-                        onClick={() => gateConnectionFlow(openPrimaryAddFlow)}
-                      >
-                        {providerSupportsPat ? "Add PAT" : t("add")}
-                      </Button>
-                      {providerId === "qoder" && (
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          onClick={() => gateConnectionFlow(() => setShowOAuthModal(true))}
-                        >
-                          Experimental OAuth
-                        </Button>
-                      )}
-                      {providerId === "codex" && (
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          icon="menu_book"
-                          onClick={() => setCodexCliGuideOpen(true)}
-                        >
-                          Codex CLI Guide
-                        </Button>
-                      )}
-                      {providerId === "codex" && (
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          icon="share"
-                          onClick={() => gateConnectionFlow(openExternalLinkFlow)}
-                        >
-                          Adicionar Externo
-                        </Button>
-                      )}
-                      {providerId === "codex" && (
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          icon="upload_file"
-                          onClick={() => gateConnectionFlow(() => setImportCodexModalOpen(true))}
-                        >
-                          {typeof t.has === "function" && t.has("importCodexAuth")
-                            ? t("importCodexAuth")
-                            : "Import auth"}
-                        </Button>
-                      )}
-                      {providerId === "claude" && (
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          icon="upload_file"
-                          onClick={() => gateConnectionFlow(() => setImportClaudeModalOpen(true))}
-                        >
-                          {typeof t.has === "function" && t.has("importClaudeAuth")
-                            ? t("importClaudeAuth")
-                            : "Import auth"}
-                        </Button>
-                      )}
-                      {providerId === "gemini-cli" && (
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          icon="upload_file"
-                          onClick={() => gateConnectionFlow(() => setImportGeminiModalOpen(true))}
-                        >
-                          {typeof t.has === "function" && t.has("importGeminiAuth")
-                            ? t("importGeminiAuth")
-                            : "Import auth"}
-                        </Button>
-                      )}
-                    </>
-                  )}
-                </>
-              ) : (
-                connections.length === 0 && (
-                  <Button
-                    size="sm"
-                    icon="add"
-                    onClick={() => gateConnectionFlow(openApiKeyAddFlow)}
-                  >
-                    {t("add")}
-                  </Button>
-                )
-              )}
-            </div>
-          </div>
+          <ConnectionsHeaderToolbar
+            providerId={providerId}
+            providerInfo={providerInfo}
+            isCompatible={isCompatible}
+            isCommandCode={isCommandCode}
+            isOAuth={isOAuth}
+            providerSupportsPat={providerSupportsPat}
+            connections={connections}
+            batchTesting={batchTesting}
+            batchRetesting={batchRetesting}
+            retestingId={retestingId}
+            distributingProxies={distributingProxies}
+            proxyConfig={proxyConfig}
+            preferClaudeCodeForUnprefixedClaudeModels={preferClaudeCodeForUnprefixedClaudeModels}
+            claudeRoutingSettingsLoaded={claudeRoutingSettingsLoaded}
+            claudeRoutingSettingsLoadError={claudeRoutingSettingsLoadError}
+            savingClaudeRoutingPreference={savingClaudeRoutingPreference}
+            handleToggleClaudeRoutingPreference={handleToggleClaudeRoutingPreference}
+            loadClaudeRoutingSettings={loadClaudeRoutingSettings}
+            codexGlobalServiceMode={codexGlobalServiceMode}
+            codexGlobalServiceModeOptions={codexGlobalServiceModeOptions}
+            codexSettingsLoaded={codexSettingsLoaded}
+            codexSettingsLoadError={codexSettingsLoadError}
+            savingCodexGlobalServiceMode={savingCodexGlobalServiceMode}
+            handleChangeCodexGlobalServiceMode={handleChangeCodexGlobalServiceMode}
+            loadCodexSettings={loadCodexSettings}
+            onSetProxyTarget={setProxyTarget}
+            handleDistributeProxies={handleDistributeProxies}
+            handleBatchTestAll={handleBatchTestAll}
+            gateConnectionFlow={gateConnectionFlow}
+            openApiKeyAddFlow={openApiKeyAddFlow}
+            openPrimaryAddFlow={openPrimaryAddFlow}
+            openExternalLinkFlow={openExternalLinkFlow}
+            handleOpenCommandCodeConnect={handleOpenCommandCodeConnect}
+            commandCodeAuthState={commandCodeAuthState}
+            onOpenOAuthModal={() => setShowOAuthModal(true)}
+            onOpenCodexCliGuide={() => setCodexCliGuideOpen(true)}
+            onOpenImportCodex={() => setImportCodexModalOpen(true)}
+            onOpenImportClaude={() => setImportClaudeModalOpen(true)}
+            onOpenImportGemini={() => setImportGeminiModalOpen(true)}
+            t={t}
+          />
 
           {connections.length === 0 ? (
             <div className="text-center py-12">
@@ -1440,548 +932,71 @@ export default function ProviderDetailPageClient() {
               )}
             </div>
           ) : (
-            (() => {
-              const sorted = [...connections].sort((a, b) => (a.priority || 0) - (b.priority || 0));
-              const hasAnyTag = sorted.some(
-                (c) => c.providerSpecificData?.tag as string | undefined
-              );
-              const allSelected = selectedIds.size === connections.length && connections.length > 0;
-              const someSelected = selectedIds.size > 0 && selectedIds.size < connections.length;
-              const bulkBusy =
-                batchUpdating !== null || batchRetesting || batchDeleting || batchTesting;
-              const bulkActions = selectedIds.size > 0 && (
-                <div className="flex flex-wrap items-center justify-end gap-2">
-                  <Button
-                    variant="secondary"
-                    size="sm"
-                    icon="toggle_on"
-                    loading={batchUpdating === "activate"}
-                    disabled={bulkBusy && batchUpdating !== "activate"}
-                    onClick={() => handleBatchSetActive(true)}
-                  >
-                    {t("batchActivateSelected")}
-                  </Button>
-                  <Button
-                    variant="secondary"
-                    size="sm"
-                    icon="toggle_off"
-                    loading={batchUpdating === "deactivate"}
-                    disabled={bulkBusy && batchUpdating !== "deactivate"}
-                    onClick={() => handleBatchSetActive(false)}
-                  >
-                    {t("batchDeactivateSelected")}
-                  </Button>
-                  <Button
-                    variant="secondary"
-                    size="sm"
-                    icon="play_arrow"
-                    loading={batchRetesting}
-                    disabled={(bulkBusy && !batchRetesting) || !!retestingId}
-                    onClick={handleBatchRetest}
-                  >
-                    {t("batchRetestSelected")}
-                  </Button>
-                  <Button
-                    variant="danger"
-                    size="sm"
-                    icon="delete"
-                    loading={batchDeleting}
-                    disabled={bulkBusy && !batchDeleting}
-                    onClick={handleBatchDeleteOpenModal}
-                  >
-                    {t("batchDeleteSelected", { count: selectedIds.size })}
-                  </Button>
-                </div>
-              );
-
-              const isHealthy = (c: ConnectionRowConnection): boolean => {
-                const s = c.testStatus;
-                return c.isActive !== false && (!s || s === "active" || s === "success");
-              };
-              const STATUS_FILTER_OPTIONS = [
-                { value: "all", label: t("filterAll", "All") },
-                { value: "active", label: t("filterActive", "Active") },
-                { value: "error", label: t("filterError", "Error") },
-                { value: "banned", label: t("filterBanned", "Banned") },
-                {
-                  value: "credits_exhausted",
-                  label: t("filterCreditsExhausted", "Credits Exhausted"),
-                },
-              ];
-              const filtered =
-                healthFilter === "all"
-                  ? sorted
-                  : sorted.filter((c) => {
-                      if (healthFilter === "active") return isHealthy(c);
-                      if (healthFilter === "error")
-                        return (
-                          !isHealthy(c) &&
-                          c.testStatus !== "banned" &&
-                          c.testStatus !== "credits_exhausted"
-                        );
-                      return c.testStatus === healthFilter;
-                    });
-
-              const totalFilteredPages = Math.max(1, Math.ceil(filtered.length / PAGE_SIZE));
-              const clampedPage = Math.min(page, totalFilteredPages - 1);
-              const pageStart = clampedPage * PAGE_SIZE;
-              const pageEnd = pageStart + PAGE_SIZE;
-
-              const filterPills = (
-                <div className="flex items-center gap-1.5 flex-wrap">
-                  {STATUS_FILTER_OPTIONS.map((opt) => (
-                    <button
-                      key={opt.value}
-                      onClick={() => {
-                        setHealthFilter(opt.value);
-                        setPage(0);
-                        setSelectedIds(new Set());
-                      }}
-                      className={`px-2.5 py-1 text-xs rounded-full font-medium transition-colors ${
-                        healthFilter === opt.value
-                          ? "bg-primary text-white"
-                          : "bg-muted/60 text-text-muted hover:bg-muted"
-                      }`}
-                    >
-                      {opt.label}
-                    </button>
-                  ))}
-                </div>
-              );
-
-              const paginationBar =
-                totalFilteredPages > 1 ? (
-                  <div className="flex items-center justify-between px-3 py-2 border-t border-border">
-                    <span className="text-xs text-text-muted">
-                      {pageStart + 1}–{Math.min(pageEnd, filtered.length)} / {filtered.length}
-                    </span>
-                    <div className="flex items-center gap-1">
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        icon="chevron_left"
-                        disabled={clampedPage === 0}
-                        onClick={() => setPage((p) => Math.max(0, p - 1))}
-                      />
-                      <span className="text-xs text-text-muted min-w-[4rem] text-center">
-                        {clampedPage + 1} / {totalFilteredPages}
-                      </span>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        icon="chevron_right"
-                        disabled={clampedPage >= totalFilteredPages - 1}
-                        onClick={() => setPage((p) => Math.min(totalFilteredPages - 1, p + 1))}
-                      />
-                    </div>
-                  </div>
-                ) : null;
-
-              if (!hasAnyTag) {
-                const pageConnections = filtered.slice(pageStart, pageEnd);
-                const allSelected =
-                  pageConnections.length > 0 && pageConnections.every((c) => selectedIds.has(c.id));
-                const someSelected = pageConnections.some((c) => selectedIds.has(c.id));
-                return (
-                  <>
-                    <div className="flex flex-col sm:flex-row items-start sm:items-center justify-between gap-2 px-3 py-2 bg-muted/50 rounded-t-lg border border-b-0 border-border">
-                      <div className="flex items-center gap-2 flex-wrap">
-                        <label className="flex items-center gap-2 cursor-pointer select-none">
-                          <input
-                            type="checkbox"
-                            checked={allSelected}
-                            ref={(el) => {
-                              if (el) el.indeterminate = someSelected;
-                            }}
-                            onChange={() => {
-                              if (allSelected) {
-                                const toRemove = new Set(pageConnections.map((c) => c.id));
-                                setSelectedIds((prev) => {
-                                  const next = new Set(prev);
-                                  for (const id of toRemove) next.delete(id);
-                                  return next;
-                                });
-                              } else {
-                                setSelectedIds((prev) => {
-                                  const next = new Set(prev);
-                                  for (const c of pageConnections) next.add(c.id);
-                                  return next;
-                                });
-                              }
-                            }}
-                            className="w-4 h-4 rounded border-border text-primary focus:ring-primary/30 cursor-pointer"
-                          />
-                          <span className="text-sm font-medium text-text-muted">
-                            {selectedIds.size > 0
-                              ? providerCountText(
-                                  t,
-                                  "selectedCount",
-                                  selectedIds.size,
-                                  "{count} selected",
-                                  "{count} selected"
-                                )
-                              : providerCountText(
-                                  t,
-                                  "accountsCount",
-                                  filtered.length,
-                                  "{count} account",
-                                  "{count} accounts"
-                                )}
-                          </span>
-                        </label>
-                        {filterPills}
-                      </div>
-
-                      {bulkActions}
-                    </div>
-                    <div className="flex flex-col divide-y divide-black/[0.03] dark:divide-white/[0.03] border border-t-0 border-border rounded-b-lg overflow-hidden">
-                      {pageConnections.length === 0 ? (
-                        <div className="px-3 py-6 text-center text-sm text-text-muted">
-                          {t("noFilteredConnections", "No connections match the current filter.")}
-                        </div>
-                      ) : (
-                        pageConnections.map((conn, index) => (
-                          <ConnectionRow
-                            key={conn.id}
-                            connection={conn}
-                            isOAuth={conn.authType === "oauth"}
-                            isClaude={providerId === "claude"}
-                            codexGlobalServiceMode={codexGlobalServiceMode}
-                            isFirst={index === 0}
-                            isLast={index === pageConnections.length - 1}
-                            isSelected={selectedIds.has(conn.id)}
-                            onToggleSelect={() => handleToggleSelectOne(conn.id)}
-                            onMoveUp={() => handleSwapPriority(conn, sorted[index - 1])}
-                            onMoveDown={() => handleSwapPriority(conn, sorted[index + 1])}
-                            onToggleActive={(isActive) =>
-                              handleUpdateConnectionStatus(conn.id, isActive)
-                            }
-                            onToggleRateLimit={(enabled) => handleToggleRateLimit(conn.id, enabled)}
-                            onToggleClaudeExtraUsage={(enabled) =>
-                              handleToggleClaudeExtraUsage(conn.id, enabled)
-                            }
-                            isCodex={providerId === "codex"}
-                            isGeminiCli={providerId === "gemini-cli"}
-                            isCcCompatible={isCcCompatible}
-                            cliproxyapiEnabled={cpaProviderEnabled}
-                            onToggleCliproxyapiMode={(enabled) =>
-                              handleToggleCliproxyapiMode(conn.id, enabled)
-                            }
-                            onToggleCodex5h={(enabled) =>
-                              handleToggleCodexLimit(conn.id, "use5h", enabled)
-                            }
-                            onToggleCodexWeekly={(enabled) =>
-                              handleToggleCodexLimit(conn.id, "useWeekly", enabled)
-                            }
-                            onRetest={() => handleRetestConnection(conn.id)}
-                            isRetesting={retestingId === conn.id}
-                            onEdit={() => {
-                              setSelectedConnection(conn);
-                              setShowEditModal(true);
-                            }}
-                            onDelete={() => handleDelete(conn.id)}
-                            onReauth={
-                              conn.authType === "oauth"
-                                ? () => gateConnectionFlow(() => setShowOAuthModal(true, conn))
-                                : undefined
-                            }
-                            onRefreshToken={
-                              conn.authType === "oauth"
-                                ? () => handleRefreshToken(conn.id)
-                                : undefined
-                            }
-                            isRefreshing={refreshingId === conn.id}
-                            onApplyCodexAuthLocal={
-                              providerId === "codex"
-                                ? () => setApplyCodexModalConnectionId(conn.id)
-                                : undefined
-                            }
-                            isApplyingCodexAuthLocal={applyingCodexAuthId === conn.id}
-                            onExportCodexAuthFile={
-                              providerId === "codex"
-                                ? () => handleExportCodexAuthFile(conn.id)
-                                : undefined
-                            }
-                            isExportingCodexAuthFile={exportingCodexAuthId === conn.id}
-                            onApplyClaudeAuthLocal={
-                              providerId === "claude"
-                                ? () => setApplyClaudeModalConnectionId(conn.id)
-                                : undefined
-                            }
-                            isApplyingClaudeAuthLocal={applyingClaudeAuthId === conn.id}
-                            onExportClaudeAuthFile={
-                              providerId === "claude"
-                                ? () => handleExportClaudeAuthFile(conn.id)
-                                : undefined
-                            }
-                            isExportingClaudeAuthFile={exportingClaudeAuthId === conn.id}
-                            onApplyGeminiAuthLocal={
-                              providerId === "gemini-cli"
-                                ? () => setApplyGeminiModalConnectionId(conn.id)
-                                : undefined
-                            }
-                            isApplyingGeminiAuthLocal={applyingGeminiAuthId === conn.id}
-                            onExportGeminiAuthFile={
-                              providerId === "gemini-cli"
-                                ? () => handleExportGeminiAuthFile(conn.id)
-                                : undefined
-                            }
-                            isExportingGeminiAuthFile={exportingGeminiAuthId === conn.id}
-                            onProxy={() =>
-                              setProxyTarget({
-                                level: "key",
-                                id: conn.id,
-                                label: pickDisplayValue(
-                                  [conn.name, conn.email],
-                                  emailsVisible,
-                                  conn.id
-                                ),
-                              })
-                            }
-                            hasProxy={!!connProxyMap[conn.id]?.proxy}
-                            proxySource={connProxyMap[conn.id]?.level || null}
-                            proxyHost={connProxyMap[conn.id]?.proxy?.host || null}
-                            proxyEnabled={readBooleanToggle(conn.proxyEnabled, true)}
-                            onToggleProxyEnabled={(enabled) =>
-                              handleToggleProxyEnabled(conn.id, enabled)
-                            }
-                            perKeyProxyEnabled={readBooleanToggle(conn.perKeyProxyEnabled, false)}
-                            onTogglePerKeyProxyEnabled={(enabled) =>
-                              handleTogglePerKeyProxyEnabled(conn.id, enabled)
-                            }
-                          />
-                        ))
-                      )}
-                    </div>
-                    {paginationBar}
-                  </>
-                );
-              }
-
-              // Build ordered tag groups: untagged first, then alphabetically
-              const groupMap = new Map<string, ConnectionRowConnection[]>();
-              for (const conn of filtered) {
-                const tag = (conn.providerSpecificData?.tag as string | undefined)?.trim() || "";
-                if (!groupMap.has(tag)) groupMap.set(tag, []);
-                groupMap.get(tag)!.push(conn);
-              }
-              const groupKeys = Array.from(groupMap.keys()).sort((a, b) => {
-                if (a === "") return -1;
-                if (b === "") return 1;
-                return compareTr(a, b);
-              });
-
-              return (
-                <>
-                  {selectedIds.size > 0 || connections.length > 0 ? (
-                    <div className="flex flex-col sm:flex-row items-start sm:items-center justify-between gap-2 px-3 py-2 bg-muted/50 rounded-t-lg border border-b-0 border-border">
-                      <div className="flex items-center gap-2 flex-wrap">
-                        <label className="flex items-center gap-2 cursor-pointer select-none">
-                          <input
-                            type="checkbox"
-                            checked={allSelected}
-                            ref={(el) => {
-                              if (el) el.indeterminate = someSelected;
-                            }}
-                            onChange={handleToggleSelectAll}
-                            className="w-4 h-4 rounded border-border text-primary focus:ring-primary/30 cursor-pointer"
-                          />
-                          <span className="text-sm font-medium text-text-muted">
-                            {selectedIds.size > 0
-                              ? providerCountText(
-                                  t,
-                                  "selectedCount",
-                                  selectedIds.size,
-                                  "{count} selected",
-                                  "{count} selected"
-                                )
-                              : providerCountText(
-                                  t,
-                                  "accountsCount",
-                                  filtered.length,
-                                  "{count} account",
-                                  "{count} accounts"
-                                )}
-                          </span>
-                        </label>
-                        {filterPills}
-                      </div>
-
-                      <div className="flex flex-wrap items-center justify-end gap-2">
-                        {/* Distribute Proxies lives in the provider toolbar (top action bar);
-                            removed the duplicate here that rendered simultaneously when nothing
-                            was selected. Per-tag groups keep their own scoped button. */}
-                        {bulkActions}
-                      </div>
-                    </div>
-                  ) : null}
-                  <div className="flex flex-col gap-0 border border-t-0 border-border rounded-b-lg overflow-hidden">
-                    {groupKeys.map((tag, gi) => {
-                      const groupConns = groupMap.get(tag)!;
-                      return (
-                        <div
-                          key={tag || "__untagged__"}
-                          className={
-                            gi > 0
-                              ? "border-t border-black/[0.06] dark:border-white/[0.06] mt-1 pt-1"
-                              : ""
-                          }
-                        >
-                          {tag && (
-                            <div className="flex items-center gap-2 px-3 pt-2 pb-1">
-                              <span className="material-symbols-outlined text-[13px] text-text-muted/50">
-                                label
-                              </span>
-                              <span className="text-[11px] font-semibold uppercase tracking-widest text-text-muted/60 select-none">
-                                {tag}
-                              </span>
-                              <div className="flex-1 h-px bg-black/[0.04] dark:bg-white/[0.04]" />
-                              <Button
-                                variant="ghost"
-                                size="sm"
-                                icon="shield"
-                                loading={distributingProxies}
-                                onClick={() => handleDistributeProxies(tag)}
-                              >
-                                Distribute Proxies
-                              </Button>
-                              <span className="text-[10px] text-text-muted/40">
-                                {groupConns.length}
-                              </span>
-                            </div>
-                          )}
-                          <div className="flex flex-col divide-y divide-black/[0.03] dark:divide-white/[0.03]">
-                            {groupConns.map((conn, index) => (
-                              <ConnectionRow
-                                key={conn.id}
-                                connection={conn}
-                                isOAuth={conn.authType === "oauth"}
-                                isClaude={providerId === "claude"}
-                                codexGlobalServiceMode={codexGlobalServiceMode}
-                                isFirst={gi === 0 && index === 0}
-                                isLast={
-                                  gi === groupKeys.length - 1 && index === groupConns.length - 1
-                                }
-                                isSelected={selectedIds.has(conn.id)}
-                                onToggleSelect={() => handleToggleSelectOne(conn.id)}
-                                onMoveUp={() =>
-                                  handleSwapPriority(conn, sorted[sorted.indexOf(conn) - 1])
-                                }
-                                onMoveDown={() =>
-                                  handleSwapPriority(conn, sorted[sorted.indexOf(conn) + 1])
-                                }
-                                onToggleActive={(isActive) =>
-                                  handleUpdateConnectionStatus(conn.id, isActive)
-                                }
-                                onToggleRateLimit={(enabled) =>
-                                  handleToggleRateLimit(conn.id, enabled)
-                                }
-                                onToggleClaudeExtraUsage={(enabled) =>
-                                  handleToggleClaudeExtraUsage(conn.id, enabled)
-                                }
-                                isCodex={providerId === "codex"}
-                                isGeminiCli={providerId === "gemini-cli"}
-                                isCcCompatible={isCcCompatible}
-                                cliproxyapiEnabled={cpaProviderEnabled}
-                                onToggleCodex5h={(enabled) =>
-                                  handleToggleCodexLimit(conn.id, "use5h", enabled)
-                                }
-                                onToggleCodexWeekly={(enabled) =>
-                                  handleToggleCodexLimit(conn.id, "useWeekly", enabled)
-                                }
-                                onRetest={() => handleRetestConnection(conn.id)}
-                                isRetesting={retestingId === conn.id}
-                                onEdit={() => {
-                                  setSelectedConnection(conn);
-                                  setShowEditModal(true);
-                                }}
-                                onDelete={() => handleDelete(conn.id)}
-                                onReauth={
-                                  conn.authType === "oauth"
-                                    ? () => gateConnectionFlow(() => setShowOAuthModal(true, conn))
-                                    : undefined
-                                }
-                                onRefreshToken={
-                                  conn.authType === "oauth"
-                                    ? () => handleRefreshToken(conn.id)
-                                    : undefined
-                                }
-                                isRefreshing={refreshingId === conn.id}
-                                onApplyCodexAuthLocal={
-                                  providerId === "codex"
-                                    ? () => setApplyCodexModalConnectionId(conn.id)
-                                    : undefined
-                                }
-                                isApplyingCodexAuthLocal={applyingCodexAuthId === conn.id}
-                                onExportCodexAuthFile={
-                                  providerId === "codex"
-                                    ? () => handleExportCodexAuthFile(conn.id)
-                                    : undefined
-                                }
-                                isExportingCodexAuthFile={exportingCodexAuthId === conn.id}
-                                onApplyClaudeAuthLocal={
-                                  providerId === "claude"
-                                    ? () => setApplyClaudeModalConnectionId(conn.id)
-                                    : undefined
-                                }
-                                isApplyingClaudeAuthLocal={applyingClaudeAuthId === conn.id}
-                                onExportClaudeAuthFile={
-                                  providerId === "claude"
-                                    ? () => handleExportClaudeAuthFile(conn.id)
-                                    : undefined
-                                }
-                                isExportingClaudeAuthFile={exportingClaudeAuthId === conn.id}
-                                onApplyGeminiAuthLocal={
-                                  providerId === "gemini-cli"
-                                    ? () => setApplyGeminiModalConnectionId(conn.id)
-                                    : undefined
-                                }
-                                isApplyingGeminiAuthLocal={applyingGeminiAuthId === conn.id}
-                                onExportGeminiAuthFile={
-                                  providerId === "gemini-cli"
-                                    ? () => handleExportGeminiAuthFile(conn.id)
-                                    : undefined
-                                }
-                                isExportingGeminiAuthFile={exportingGeminiAuthId === conn.id}
-                                onProxy={() =>
-                                  setProxyTarget({
-                                    level: "key",
-                                    id: conn.id,
-                                    label: pickDisplayValue(
-                                      [conn.name, conn.email],
-                                      emailsVisible,
-                                      conn.id
-                                    ),
-                                  })
-                                }
-                                hasProxy={!!connProxyMap[conn.id]?.proxy}
-                                proxySource={connProxyMap[conn.id]?.level || null}
-                                proxyHost={connProxyMap[conn.id]?.proxy?.host || null}
-                                proxyEnabled={readBooleanToggle(conn.proxyEnabled, true)}
-                                onToggleProxyEnabled={(enabled) =>
-                                  handleToggleProxyEnabled(conn.id, enabled)
-                                }
-                                perKeyProxyEnabled={readBooleanToggle(
-                                  conn.perKeyProxyEnabled,
-                                  false
-                                )}
-                                onTogglePerKeyProxyEnabled={(enabled) =>
-                                  handleTogglePerKeyProxyEnabled(conn.id, enabled)
-                                }
-                              />
-                            ))}
-                          </div>
-                        </div>
-                      );
-                    })}
-                  </div>
-                </>
-              );
-            })()
+            <ConnectionsListPanel
+              connections={connections}
+              providerId={providerId}
+              isCcCompatible={isCcCompatible}
+              isOAuth={isOAuth}
+              codexGlobalServiceMode={codexGlobalServiceMode}
+              selectedIds={selectedIds}
+              batchUpdating={batchUpdating}
+              batchRetesting={batchRetesting}
+              batchDeleting={batchDeleting}
+              batchTesting={batchTesting}
+              retestingId={retestingId}
+              refreshingId={refreshingId}
+              distributingProxies={distributingProxies}
+              healthFilter={healthFilter}
+              page={page}
+              PAGE_SIZE={PAGE_SIZE}
+              connProxyMap={connProxyMap}
+              proxyConfig={proxyConfig}
+              applyingCodexAuthId={applyingCodexAuthId}
+              exportingCodexAuthId={exportingCodexAuthId}
+              applyingClaudeAuthId={applyingClaudeAuthId}
+              exportingClaudeAuthId={exportingClaudeAuthId}
+              applyingGeminiAuthId={applyingGeminiAuthId}
+              exportingGeminiAuthId={exportingGeminiAuthId}
+              emailsVisible={emailsVisible}
+              setSelectedIds={setSelectedIds}
+              setPage={setPage}
+              setHealthFilter={setHealthFilter}
+              handleDelete={handleDelete}
+              handleUpdateConnectionStatus={handleUpdateConnectionStatus}
+              handleToggleRateLimit={handleToggleRateLimit}
+              handleToggleClaudeExtraUsage={handleToggleClaudeExtraUsage}
+              handleToggleCliproxyapiMode={handleToggleCliproxyapiMode}
+              handleToggleCodexLimit={handleToggleCodexLimit}
+              handleToggleProxyEnabled={handleToggleProxyEnabled}
+              handleTogglePerKeyProxyEnabled={handleTogglePerKeyProxyEnabled}
+              handleRetestConnection={handleRetestConnection}
+              handleRefreshToken={handleRefreshToken}
+              handleSwapPriority={handleSwapPriority}
+              handleBatchSetActive={handleBatchSetActive}
+              handleBatchDeleteOpenModal={handleBatchDeleteOpenModal}
+              handleBatchRetest={handleBatchRetest}
+              handleToggleSelectOne={handleToggleSelectOne}
+              handleToggleSelectAll={handleToggleSelectAll}
+              handleDistributeProxies={handleDistributeProxies}
+              cpaProviderEnabled={cpaProviderEnabled}
+              onOpenEditModal={(conn) => {
+                setSelectedConnection(conn);
+                setShowEditModal(true);
+              }}
+              onOpenOAuth={(conn) => gateConnectionFlow(() => setShowOAuthModal(true, conn))}
+              onSetProxyTarget={setProxyTarget}
+              onOpenApplyCodexModal={setApplyCodexModalConnectionId}
+              onExportCodexAuthFile={handleExportCodexAuthFile}
+              onOpenApplyClaudeModal={setApplyClaudeModalConnectionId}
+              onExportClaudeAuthFile={handleExportClaudeAuthFile}
+              onOpenApplyGeminiModal={setApplyGeminiModalConnectionId}
+              onExportGeminiAuthFile={handleExportGeminiAuthFile}
+              gateConnectionFlow={gateConnectionFlow}
+              t={t}
+            />
           )}
         </Card>
       )}
-
       {isUpstreamProxyProvider && (
         <Card>
           <div className="flex flex-col gap-3">
@@ -2309,97 +1324,14 @@ export default function ProviderDetailPageClient() {
         />
       )}
       {/* Batch Test Results Modal */}
-      {batchTestResults && (
-        <div
-          className="fixed inset-0 z-50 flex items-start justify-center pt-[10vh]"
-          onClick={() => setBatchTestResults(null)}
-        >
-          <div className="absolute inset-0 bg-black/60 backdrop-blur-sm" />
-          <div
-            className="relative bg-bg-primary border border-border rounded-xl w-full max-w-[600px] max-h-[80vh] overflow-y-auto shadow-2xl"
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div className="sticky top-0 z-10 flex items-center justify-between px-5 py-3 border-b border-border bg-bg-primary/95 backdrop-blur-sm rounded-t-xl">
-              <h3 className="font-semibold">{t("testResults")}</h3>
-              <button
-                onClick={() => setBatchTestResults(null)}
-                className="p-1 rounded-lg hover:bg-bg-subtle text-text-muted hover:text-text-primary transition-colors"
-                aria-label={t("close")}
-              >
-                <span className="material-symbols-outlined text-lg">close</span>
-              </button>
-            </div>
-            <div className="p-5">
-              {batchTestResults.error &&
-              (!batchTestResults.results || batchTestResults.results.length === 0) ? (
-                <div className="text-center py-6">
-                  <span className="material-symbols-outlined text-red-500 text-[32px] mb-2 block">
-                    error
-                  </span>
-                  <p className="text-sm text-red-400">{String(batchTestResults.error)}</p>
-                </div>
-              ) : (
-                <div className="flex flex-col gap-3">
-                  {batchTestResults.summary && (
-                    <div className="flex items-center gap-3 text-xs mb-1">
-                      <span className="text-text-muted">{providerInfo?.name || providerId}</span>
-                      <span className="px-2 py-0.5 rounded bg-emerald-500/15 text-emerald-400 font-medium">
-                        {t("passedCount", { count: batchTestResults.summary.passed })}
-                      </span>
-                      {batchTestResults.summary.failed > 0 && (
-                        <span className="px-2 py-0.5 rounded bg-red-500/15 text-red-400 font-medium">
-                          {t("failedCount", { count: batchTestResults.summary.failed })}
-                        </span>
-                      )}
-                      <span className="text-text-muted ml-auto">
-                        {t("testedCount", { count: batchTestResults.summary.total })}
-                      </span>
-                    </div>
-                  )}
-                  {(batchTestResults.results || []).map((r: any, i: number) => (
-                    <div
-                      key={r.connectionId || i}
-                      className="flex items-center gap-2 text-xs px-3 py-2 rounded-lg bg-black/[0.03] dark:bg-white/[0.03]"
-                    >
-                      <span
-                        className={`material-symbols-outlined text-[16px] ${
-                          r.valid ? "text-emerald-500" : "text-red-500"
-                        }`}
-                      >
-                        {r.valid ? "check_circle" : "error"}
-                      </span>
-                      <div className="flex-1 min-w-0">
-                        <span className="font-medium">
-                          {pickDisplayValue([r.connectionName], emailsVisible, r.connectionName)}
-                        </span>
-                      </div>
-                      {r.latencyMs !== undefined && (
-                        <span className="text-text-muted font-mono tabular-nums">
-                          {t("millisecondsAbbr", { value: r.latencyMs })}
-                        </span>
-                      )}
-                      <span
-                        className={`text-[10px] uppercase font-bold px-1.5 py-0.5 rounded ${
-                          r.valid
-                            ? "bg-emerald-500/15 text-emerald-400"
-                            : "bg-red-500/15 text-red-400"
-                        }`}
-                      >
-                        {r.valid ? t("okShort") : r.diagnosis?.type || t("errorShort")}
-                      </span>
-                    </div>
-                  ))}
-                  {(!batchTestResults.results || batchTestResults.results.length === 0) && (
-                    <div className="text-center py-4 text-text-muted text-sm">
-                      {t("noActiveConnectionsInGroup")}
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
+      <BatchTestResultsModal
+        batchTestResults={batchTestResults}
+        providerInfo={providerInfo}
+        providerId={providerId}
+        emailsVisible={emailsVisible}
+        onClose={() => setBatchTestResults(null)}
+        t={t}
+      />
       {/* Proxy Config Modal */}
       {proxyTarget && (
         <ProxyConfigModal
@@ -2427,119 +1359,10 @@ export default function ProviderDetailPageClient() {
 
       {/* Adapta Web — Tutorial Modal */}
       {providerId === "adapta-web" && (
-        <Modal
+        <AdaptaTutorialModal
           isOpen={showTutorialModal}
           onClose={() => setShowTutorialModal(false)}
-          title="Como conectar o Adapta Web"
-          size="md"
-        >
-          <div className="flex flex-col gap-5 text-sm">
-            <p className="text-text-muted">
-              O Adapta usa autenticação via Clerk. O token{" "}
-              <code className="bg-surface-2 px-1 rounded font-mono text-xs">__client</code> é um JWT
-              de longa duração que permite renovar sessões automaticamente.
-            </p>
-
-            <ol className="flex flex-col gap-4 list-none">
-              <li className="flex gap-3">
-                <span className="flex-none w-6 h-6 rounded-full bg-primary text-white text-xs font-bold flex items-center justify-center">
-                  1
-                </span>
-                <div>
-                  <p className="font-medium">Acesse o chat do Adapta</p>
-                  <p className="text-text-muted mt-0.5">
-                    Abra{" "}
-                    <a
-                      href="https://agent.adapta.one/agentic-chat"
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="underline text-primary"
-                    >
-                      agent.adapta.one/agentic-chat
-                    </a>{" "}
-                    e faça login com sua conta Gold ou Business.
-                  </p>
-                </div>
-              </li>
-
-              <li className="flex gap-3">
-                <span className="flex-none w-6 h-6 rounded-full bg-primary text-white text-xs font-bold flex items-center justify-center">
-                  2
-                </span>
-                <div>
-                  <p className="font-medium">Abra o DevTools</p>
-                  <p className="text-text-muted mt-0.5">
-                    Pressione{" "}
-                    <kbd className="bg-surface-2 px-1.5 py-0.5 rounded text-xs font-mono">F12</kbd>{" "}
-                    ou{" "}
-                    <kbd className="bg-surface-2 px-1.5 py-0.5 rounded text-xs font-mono">
-                      Cmd+Option+I
-                    </kbd>{" "}
-                    para abrir as Ferramentas do Desenvolvedor.
-                  </p>
-                </div>
-              </li>
-
-              <li className="flex gap-3">
-                <span className="flex-none w-6 h-6 rounded-full bg-primary text-white text-xs font-bold flex items-center justify-center">
-                  3
-                </span>
-                <div>
-                  <p className="font-medium">Vá em Application → Cookies</p>
-                  <p className="text-text-muted mt-0.5">
-                    Na aba <strong>Application</strong> (Chrome/Edge) ou <strong>Storage</strong>{" "}
-                    (Firefox), expanda <strong>Cookies</strong> e clique em{" "}
-                    <code className="bg-surface-2 px-1 rounded font-mono text-xs">
-                      .clerk.agent.adapta.one
-                    </code>
-                    .
-                  </p>
-                </div>
-              </li>
-
-              <li className="flex gap-3">
-                <span className="flex-none w-6 h-6 rounded-full bg-primary text-white text-xs font-bold flex items-center justify-center">
-                  4
-                </span>
-                <div>
-                  <p className="font-medium">
-                    Copie o valor do cookie{" "}
-                    <code className="bg-surface-2 px-1 rounded font-mono text-xs">__client</code>
-                  </p>
-                  <p className="text-text-muted mt-0.5">
-                    Localize o cookie chamado{" "}
-                    <code className="bg-surface-2 px-1 rounded font-mono text-xs">__client</code> na
-                    lista. Clique nele e copie o conteúdo da coluna <strong>Value</strong> — começa
-                    com <code className="bg-surface-2 px-1 rounded font-mono text-xs">eyJ…</code>.
-                  </p>
-                </div>
-              </li>
-
-              <li className="flex gap-3">
-                <span className="flex-none w-6 h-6 rounded-full bg-primary text-white text-xs font-bold flex items-center justify-center">
-                  5
-                </span>
-                <div>
-                  <p className="font-medium">Cole aqui e salve</p>
-                  <p className="text-text-muted mt-0.5">
-                    Clique em <strong>Add Connection</strong>, cole o valor do{" "}
-                    <code className="bg-surface-2 px-1 rounded font-mono text-xs">__client</code> no
-                    campo de API Key e salve. O OmniRoute renovará a sessão automaticamente.
-                  </p>
-                </div>
-              </li>
-            </ol>
-
-            <div
-              className="rounded-lg p-3 text-xs text-text-muted"
-              style={{ backgroundColor: "rgba(110,58,211,0.08)", borderLeft: "3px solid #6E3AD3" }}
-            >
-              <strong>Dica:</strong> O cookie <code className="font-mono">__client</code> tem
-              validade longa (meses). Só será necessário renová-lo se você sair da conta ou o Adapta
-              invalidar a sessão.
-            </div>
-          </div>
-        </Modal>
+        />
       )}
     </div>
   );
